refactor(materials): deduplicate neo-Hookean Pence-Gou formulations, simplify material dispatch
The Wa/Wb/Wc (and their plastic counterparts) formulation files shared
large amounts of byte-identical boilerplate: constructor,
assignCurrentStateVars, getResult, and for the plastic variants the
entire J2 return-mapping Newton solve. Extract these into
NeoHookeanPenceGouBaseMaterial and NeoHookeanPenceGouPlasticBaseMaterial;
each formulation file now only implements its law-specific energy/stress
functions.

Also:
- replace the if/elif/strCaseCmp import chain in materiallibrary.py
  with a dict-based dispatch table
- make getDensity() and the "uniaxial stress not possible" behavior
  concrete defaults on the base classes instead of duplicated overrides
- drop two dead commented-out imports

No numerical/behavioral changes; verified against the full
testfiles/edelweiss-only/ suite (0 failures), including all six
neo-Hookean elastic/plastic Wa/Wb/Wc variants.
