+++ edelweissfe/config/materiallibrary.py
@@ -26,7 +26,42 @@
 #  the top level directory of EdelweissFE.
 #  ---------------------------------------------------------------------
 
+from importlib import import_module
+
 from edelweissfe.utils.misc import strCaseCmp
+
+# materialName (lowercase) -> (module path, class name)
+_EDELWEISS_MATERIALS = {
+    "linearelastic": ("edelweissfe.materials.linearelastic.linearelastic", "LinearElasticMaterial"),
+    "vonmises": ("edelweissfe.materials.vonmises.vonmises", "VonMisesMaterial"),
+    "neohookewa": ("edelweissfe.materials.neohooke.neohookepencegouformulationa", "NeoHookeanWaMaterial"),
+    "neohookewb": ("edelweissfe.materials.neohooke.neohookepencegouformulationb", "NeoHookeanWbMaterial"),
+    "neohookewc": ("edelweissfe.materials.neohooke.neohookepencegouformulationc", "NeoHookeanWcMaterial"),
+    "hyperelasticadvanced": (
+        "edelweissfe.materials.hyperelasticadvanced.hyperelasticadvanced",
+        "HyperelasticAdvancedMaterial",
+    ),
+    "hyperelasticadvancedi2extended": (
+        "edelweissfe.materials.hyperelasticadvanced.hyperelasticadvancedi2extended",
+        "HyperelasticAdvancedI2ExtendedMaterial",
+    ),
+    "neohookewaplastic": (
+        "edelweissfe.materials.neohookeplastic.neohookepencegouformulationaplastic",
+        "NeoHookeanWaPlasticMaterial",
+    ),
+    "neohookewbplastic": (
+        "edelweissfe.materials.neohookeplastic.neohookepencegouformulationbplastic",
+        "NeoHookeanWbPlasticMaterial",
+    ),
+    "neohookewcplastic": (
+        "edelweissfe.materials.neohookeplastic.neohookepencegouformulationcplastic",
+        "NeoHookeanWcPlasticMaterial",
+    ),
+    "hyperplasticadvanced": (
+        "edelweissfe.materials.hyperplasticadvanced.hyperplasticadvanced",
+        "HyperplasticAdvancedMaterial",
+    ),
+}
 
 
 def getMaterialClass(materialName: str, provider: str = None) -> type:
@@ -53,71 +88,8 @@ def getMaterialClass(materialName: str, provider: str = None) -> type:
         return None
 
     if strCaseCmp(provider, "edelweiss"):
-        if strCaseCmp(materialName, "linearelastic"):
-            from edelweissfe.materials.linearelastic.linearelastic import (
-                LinearElasticMaterial,
-            )
-
-            material = LinearElasticMaterial
-        elif strCaseCmp(materialName, "vonmises"):
-            from edelweissfe.materials.vonmises.vonmises import VonMisesMaterial
-
-            material = VonMisesMaterial
-        elif strCaseCmp(materialName, "neohookewa"):
-            from edelweissfe.materials.neohooke.neohookepencegouformulationa import (
-                NeoHookeanWaMaterial,
-            )
-
-            material = NeoHookeanWaMaterial
-        elif strCaseCmp(materialName, "neohookewb"):
-            from edelweissfe.materials.neohooke.neohookepencegouformulationb import (
-                NeoHookeanWbMaterial,
-            )
-
-            material = NeoHookeanWbMaterial
-        elif strCaseCmp(materialName, "neohookewc"):
-            from edelweissfe.materials.neohooke.neohookepencegouformulationc import (
-                NeoHookeanWcMaterial,
-            )
-
-            material = NeoHookeanWcMaterial
-        elif strCaseCmp(materialName, "hyperelasticadvanced"):
-            from edelweissfe.materials.hyperelasticadvanced.hyperelasticadvanced import (
-                HyperelasticAdvancedMaterial,
-            )
-
-            material = HyperelasticAdvancedMaterial
-        elif strCaseCmp(materialName, "hyperelasticadvancedi2extended"):
-            from edelweissfe.materials.hyperelasticadvanced.hyperelasticadvancedi2extended import (
-                HyperelasticAdvancedI2ExtendedMaterial,
-            )
-
-            material = HyperelasticAdvancedI2ExtendedMaterial
-        elif strCaseCmp(materialName, "neohookewaplastic"):
-            from edelweissfe.materials.neohookeplastic.neohookepencegouformulationaplastic import (
-                NeoHookeanWaPlasticMaterial,
-            )
-
-            material = NeoHookeanWaPlasticMaterial
-        elif strCaseCmp(materialName, "neohookewbplastic"):
-            from edelweissfe.materials.neohookeplastic.neohookepencegouformulationbplastic import (
-                NeoHookeanWbPlasticMaterial,
-            )
-
-            material = NeoHookeanWbPlasticMaterial
-        elif strCaseCmp(materialName, "neohookewcplastic"):
-            from edelweissfe.materials.neohookeplastic.neohookepencegouformulationcplastic import (
-                NeoHookeanWcPlasticMaterial,
-            )
-
-            material = NeoHookeanWcPlasticMaterial
-        elif strCaseCmp(materialName, "hyperplasticadvanced"):
-            from edelweissfe.materials.hyperplasticadvanced.hyperplasticadvanced import (
-                HyperplasticAdvancedMaterial,
-            )
-
-            material = HyperplasticAdvancedMaterial
-        else:
+        modulePath, className = _EDELWEISS_MATERIALS.get(materialName.lower(), (None, None))
+        if modulePath is None:
             raise Exception("This material type doesn't exist (yet). Chosen material was: " + materialName)
 
-        return material
+        return getattr(import_module(modulePath), className)
